A bunch of renaming, monadic hooks, and some refactoring
