+++ model/shared/src/main/scala/explore/model/TileIds.scala
@@ -65,5 +65,5 @@ enum ConstraintTabTileIds(val id: NonEmptyString):
 enum OverviewTabTileIds(val id: NonEmptyString):
   case WarningsAndErrorsId extends OverviewTabTileIds("warningsAndErrors".refined)
   case GroupWarningsId     extends OverviewTabTileIds("groupWarnings".refined)
-  case ObsAttachmentsId    extends OverviewTabTileIds("obsAttachments".refined)
+  case AttachmentsId       extends OverviewTabTileIds("attachments".refined)
   case DescriptionId       extends OverviewTabTileIds("programDescription".refined)
